migrated models to ts
